fix: build process without using cbor-diag (rust)

+++ public/imzero/generate.sh
@@ -3,6 +3,16 @@ set -ev
 set -o pipefail
 here=$(dirname "$(readlink -f "$BASH_SOURCE")")
 cd "$here"
+
+if [ -z "${IMZERO_CPP_BINDING_DIR}" ]; then
+   echo "IMZERO_CPP_BINDING_DIR env variable is not set"
+   exit 1
+fi
+if [ -z "${IMZERO_DOXYGEN}" ]; then
+   echo "IMZERO_DOXYGEN env variable is not set"
+   exit 1
+fi
+
 find . -type f -name "*.out.go" -delete
 ./build.sh
 ./imgui/doxygen/generate.sh
